+++ backend/src/scripts/clearData/clearUsers.js
@@ -3,12 +3,13 @@
  * @file /docman/backend/src/scripts/clearData/clearUsers.js
  * @description Script to clear all users from the database
  * @author Richard Bakos
- * @version 2.2.1
+ * @version 2.2.2
  * @license UNLICENSED
  */
 import mongoose from 'mongoose';
 import { connectDB } from '../../config/db.js';
 import User from '../../models/User.js';
+import { assertSafeClear } from '../lib/assertSafeClear.js';
 
 /**
  * Clear all users from the database
@@ -18,6 +19,7 @@ import User from '../../models/User.js';
  */
 const clearUsers = async () => {
     try {
+        assertSafeClear('clearUsers');
         await connectDB();
         
         const result = await User.deleteMany({});
